+++ src/Conn.h
@@ -83,9 +83,9 @@ public:
 	// be recorded, otherwise just up through the transport header.
 	// Both are assumed set to true when called.
 	void NextPacket(double t, int is_orig,
-			const IP_Hdr* ip, int len, int caplen,
+			const IP_Hdr* ip, uint64_t len, uint64_t caplen,
 			const u_char*& data,
-			int& record_packet, int& record_content,
+			unsigned int& record_packet, unsigned int& record_content,
 			// arguments for reproducing packets
 			const Packet *pkt);
