+++ src/analyzer/protocol/dhcp/DHCP.cc
@@ -22,7 +22,7 @@ void DHCP_Analyzer::Done()
 	}
 
 void DHCP_Analyzer::DeliverPacket(uint64_t len, const u_char* data,
-			bool orig, uint64_t seq, const IP_Hdr* ip, int caplen)
+			bool orig, uint64_t seq, const IP_Hdr* ip, uint64_t caplen)
 	{
 	Analyzer::DeliverPacket(len, data, orig, seq, ip, caplen);
