+++ src/analyzer/protocol/dhcp/DHCP.h
@@ -14,7 +14,7 @@ public:
 
 	void Done() override;
 	void DeliverPacket(uint64_t len, const u_char* data, bool orig,
-	                   uint64_t seq, const IP_Hdr* ip, int caplen) override;
+	                   uint64_t seq, const IP_Hdr* ip, uint64_t caplen) override;
 
 	static analyzer::Analyzer* Instantiate(Connection* conn)
 		{ return new DHCP_Analyzer(conn); }
