+++ src/analyzer/protocol/dns/DNS.h
@@ -353,7 +353,7 @@ public:
 	~DNS_Analyzer() override;
 
 	void DeliverPacket(uint64_t len, const u_char* data, bool orig,
-					uint64_t seq, const IP_Hdr* ip, int caplen) override;
+					uint64_t seq, const IP_Hdr* ip, uint64_t caplen) override;
 
 	void Init() override;
 	void Done() override;
