+++ src/analyzer/protocol/ident/Ident.h
@@ -13,7 +13,7 @@ public:
 	explicit Ident_Analyzer(Connection* conn);
 	void Done() override;
 
-	void DeliverStream(int length, const u_char* data, bool is_orig) override;
+	void DeliverStream(uint64_t length, const u_char* data, bool is_orig) override;
 
 	static analyzer::Analyzer* Instantiate(Connection* conn)
 		{ return new Ident_Analyzer(conn); }
