+++ src/analyzer/protocol/radius/RADIUS.cc
@@ -25,7 +25,7 @@ void RADIUS_Analyzer::Done()
 	}
 
 void RADIUS_Analyzer::DeliverPacket(uint64_t len, const u_char* data,
-				    bool orig, uint64_t seq, const IP_Hdr* ip, int caplen)
+				    bool orig, uint64_t seq, const IP_Hdr* ip, uint64_t caplen)
 	{
 	Analyzer::DeliverPacket(len, data, orig, seq, ip, caplen);
