+++ src/analyzer/protocol/sip/SIP.cc
@@ -21,7 +21,7 @@ void SIP_Analyzer::Done()
 	}
 
 void SIP_Analyzer::DeliverPacket(uint64_t len, const u_char* data, bool orig,
-								 uint64_t seq, const IP_Hdr* ip, int caplen)
+								 uint64_t seq, const IP_Hdr* ip, uint64_t caplen)
 	{
 	bool real_orig = true;
 	if ( len > 6 && data[0] == 'S' && data[1] == 'I' && data[2] == 'P' && data[3] == '/' )
